Create Electronics Store Database and Stats
Here I created a table for an electronics store labeling the products they sell, the price it costs, the quantity in stock, the price it sells for, and how many they sold.  
I pulled all products to get a visual of the data. I then decided to go into specifics regarding the sales to help the customers better understand what is selling and what is not. I made several queries showing the list of products from least to the greatest amount sold. Queries show the item that sold the most and the item that cost the most. I thought it helpful to have the total cost of an item and the total amount made from an item using Algebraic expressions inside the statement.